Add position indexing and separate shortnaming.
1) The references in the Statutes index can appear in the wrong order or
2) can involve formatting like small caps (\textsc) which cannot be used as
the internal TeX handle. This change solves both problems. A new
optional argument is added to \newstatute. The argument can take the
following forms:

\newstatute[AA@]{MCL}{} - put the MCL statute at place AA in the index.

\newstatute[am1]{\textsc{U.S. Const.} amend. I} - use am1 as the handle
for the first ammendment. (The \textsc prevents the name from being used
as a handle.)

\newstatute[1@am1]{\textsc{U.S. Const.} amend. I} - use am1 as the
handle for the first ammendment, put it at position 1 in the index.